+++ changelog.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.1.2] - 2019-07-30
+### Added
+- Allow to free buy with a negative balance
+- Add parameter `$allowZero` to method `canWithdraw`
+
+### Fixed
+- method canWithdraw, with a negative price, almost always true
+
 ## [3.1.1] - 2019-07-29
 ### Added 
 - Add getCurrencyAttribute
@@ -354,7 +362,8 @@ The operation is now executed in the transaction and updates the new `refund` fi
 - Exceptions: AmountInvalid, BalanceIsEmpty.
 - Models: Transfer, Transaction.
 
-[Unreleased]: https://github.com/bavix/laravel-wallet/compare/3.1.1...HEAD
+[Unreleased]: https://github.com/bavix/laravel-wallet/compare/3.1.2...HEAD
+[3.1.2]: https://github.com/bavix/laravel-wallet/compare/3.1.1...3.1.2
 [3.1.1]: https://github.com/bavix/laravel-wallet/compare/3.1.0...3.1.1
 [3.1.0]: https://github.com/bavix/laravel-wallet/compare/3.0.4...3.1.0
 [3.0.4]: https://github.com/bavix/laravel-wallet/compare/3.0.3...3.0.4